Refactor/clean up buttons/widget lib
