+++ lib/theme.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 ThemeData concordiumTheme() {
   const textDisplaySmall = TextStyle(fontSize: 24, fontWeight: FontWeight.bold);
@@ -24,3 +25,29 @@ ThemeData concordiumTheme() {
     ),
   );
 }
+
+// TODO use fonts as part of theme
+
+final bodyL = GoogleFonts.getFont(
+  "IBM Plex Sans",
+  fontWeight: FontWeight.w400,
+  fontSize: 16,
+);
+
+final bodyS = GoogleFonts.getFont(
+  "IBM Plex Sans",
+  fontWeight: FontWeight.w400,
+  fontSize: 12,
+);
+
+final heading2 = GoogleFonts.getFont(
+  "IBM Plex Sans",
+  fontWeight: FontWeight.w500,
+  fontSize: 24,
+);
+
+final heading5 = GoogleFonts.getFont(
+  "IBM Plex Sans",
+  fontWeight: FontWeight.w500,
+  fontSize: 16,
+);
